requirements.txt
google-generativeai==0.8.4
streamlit==1.42.0
python-dotenv==1.0.1
